Refactor space_x => rockets
