+++ lib/main.dart
@@ -14,7 +14,7 @@ import 'ui/news/screen.dart';
 import 'ui/planets/details/edit.dart';
 import 'ui/planets/screen.dart';
 import 'ui/settings.dart';
-import 'ui/space_x/screen.dart';
+import 'ui/rockets/screen.dart';
 
 final AppModel model = AppModel();
