+++ lib/ui/rockets/details/capsule.dart
