+++ lib/ui/rockets/details/roadster.dart
