+++ lib/ui/rockets/details/rocket.dart
