+++ lib/ui/rockets/details/ship.dart
