+++ lib/ui/rockets/dialog/capsule.dart
